+++ app/code/local/Stuntcoders/Slideshow/Block/Slideshow.php
@@ -2,43 +2,56 @@
 
 class Stuntcoders_Slideshow_Block_Slideshow extends Mage_Core_Block_Template
 {
-    public function getImages()
+    public function getImagesCollection()
     {
-        if (!$this->hasLoadedImages()) {
-            if (!$this->getSlideshow()) {
-                $this->_loadSlideshow();
-            }
-
-            $this->setLoadedImages($this->getSlideshow() ? $this->getSlideshow()->getImages(): array());
-        }
-
-        return $this->getLoadedImages();
+        return $this->getSlideshow() ? $this->getSlideshow()->getImagesCollection(): array();
     }
 
     public function isEnabled()
+    {
+        return (bool) $this->getSlideshow() ? $this->getSlideshow()->getIsEnabled() : false;
+    }
+
+    public function getConfig()
+    {
+        return $this->getSlideshow() ? $this->getSlideshow()->getConfig() : '';
+    }
+
+    public function getImageUrl($image)
+    {
+        return Mage::helper('stuntcoders_slideshow')->getImageUrl($image->getImage());
+    }
+
+    public function getSlideshow()
     {
         if (!$this->hasSlideshow()) {
             $this->_loadSlideshow();
         }
 
-        return (bool) $this->getSlideshow()->getIsEnabled();
+        return $this->getData('slideshow');
     }
 
     protected function _loadSlideshow()
     {
         if ($this->hasSlideshowCode()) {
-            $this->setId(Mage::getModel('stuntcoders_slideshow/slideshow')->getIdByCode($this->getSlideshowCode()));
+            $this->setSlideshowId(
+                Mage::getModel('stuntcoders_slideshow/slideshow')->getIdByCode($this->getSlideshowCode())
+            );
         }
 
-        if ($this->hasId()) {
-            $this->setSlideshow(Mage::getModel('stuntcoders_slideshow/slideshow')->load($this->getId()));
+        if ($this->getSlideshowId()) {
+            $this->setSlideshow(Mage::getModel('stuntcoders_slideshow/slideshow')->load($this->getSlideshowId()));
         }
 
         return $this;
     }
 
-    public function getImagePath($image)
+    protected function _toHtml()
     {
-        return Mage::getBaseUrl('media') . $image['image'];
+        if (!$this->getTemplate()) {
+            $this->setTemplate('stuntcoders/slideshow/slideshow.phtml');
+        }
+        $html = $this->renderView();
+        return $html;
     }
 }