+++ app/code/local/Stuntcoders/Slideshow/Model/Resource/Slideshow/Image/Collection.php
@@ -7,4 +7,10 @@ class Stuntcoders_Slideshow_Model_Resource_Slideshow_Image_Collection
     {
         $this->_init('stuntcoders_slideshow/slideshow_image');
     }
+
+    public function addSlideshowFilter($slideshowId)
+    {
+        $this->getSelect()->where('main_table.slideshow_id in (?)', $slideshowId);
+        return $this;
+    }
 }