+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.2.0 - 2015-09-30
+### ADDED:
+- Implement editing images
+- Slick.js
+- Slideshow options
+
+### CHANGED:
+- Rename slideshow config column
+- Fix adding images to slideshow
+
 ## 0.1.0 - 2015-09-24
 ### ADDED:
 - Changelog file